UNTESTED new config for MultiQC.
Should trim off _R1 and _R2 from sample names, meaning that we get one row per sample in the general stats table.

Also added a comment at the top of the report saying that it came from the NGI-RNAseq pipeline, with a link to the docs.

NB: Not tested with Nextflow
NB: May break when used with funk account, which has an alias for multiqc with a -c config file. Not sure what happens when you use -c twice.