Fix gtf/fasta flow when additional reference is used.
The pipeline used to fail with `--skip_rsem` and gzipped reference data.
The reason for this was that the entire code block for decompressing
GTF and FASTA files was only executed because of RSEM requiring
them for building its reference.

By using `--skip_rsem` the references were not extracted
and the process combining normal+additional references
failed with a missing channel.

This commit ensures that the references are unzipped whenever
an additional fasta file is used. I added `--skip_rsem` as
additional test case to the CI script.