Embed VST assay into DESeq object.
Variance-stabilised matrix is now embedded in the single DESeqDataSet
object, so that any user-based re-organisation of the object is
mirrored in the vst matrix automatically.

Also parametrise number of gene taken into account for PCA - author
prefers all genes, as it is important to detect synchronised but
individually small-amplitude signals at this stage, to diagnose batch
effect problems.  But introduce a parameter 'ntop' so this can be
over-ridden by user.

Detect which PCs are explained most by the user-supplied
'condition'. Plot the most 'experimentally interesting' PCs, as an addition
to the 'blind' PC1 and PC2.