test(star_salmon): Rewrite with pytest-workflow
This one starts to get a little clever. Here's some of the highlights:
- All of the tests have the main job name as a tag(`star_salmon`) in
this cause. This will allow for the job to select star_salmon from
`star_rsem`. I'm debating turning these into two separate tags so that
if a change to the star module was made we can run all the tests related
to that.
- The parameter tag then allows for the specific parameter being tested
to be selected for.
- We could just run `--tag star_salmon` by itself but the
matrix.parameters will make the CI jobs run quicker.
- Files and md5 hashes will come later.
- Had to cut off the `--` before each tag because it broke the regex occasionally