Fix DESeq2 script to accept new sample sheet format
We want the R script to produce a monochrome scatter plot of the
samples, as we no longer have the grouping information explicitly in
the sample sheet.  To aid interpretation in the case where design
information is implicitly annotated in the sample names via underscore
field separators, we produce an additional plot breaking each PC up by
samples corresponding to these inferred experimental factors.

Documentation also updated to reflect the monochrome output.