Fix file path composition
Path names should be concatenated using the `resolve` API. The string interpolation should never be used with file paths, because it's not portable since it cause the loosing of the file protocol (due to the underlying java API). 

For example if the path is a S3 path eg. `file = "s3://bucket/foo/bar.txt"` the `"$file"` would return `/bucket/foo/bar.txt`.