Update modules.config
The function call of featurecounts expects for parameter -g the attribute type in the GTF annotation which is 'gene_id' by default. The way featurecounts was called in this version 3.5 and at least in version 3.4 provides the biotype instead of the expected  attribute type. Thus the results of featurecounts show the counts according to the biotypes instead of the gene ids.